reverted demo changes

+++ Data/Data/ArtStGoAmlReportsPerCreator.cs
@@ -9,6 +9,6 @@ namespace Data.Data
     public class ArtStGoAmlReportsPerCreator
     {
         public string? CREATED_BY { get; set; }
-        public int? NUMBER_OF_REPORTS { get; set; }
+        public decimal? NUMBER_OF_REPORTS { get; set; }
     }
 }
